Fixed all Clang and GCC warnings and errors

+++ Modules/Resource/source/Resources/BinaryFile.cpp
@@ -31,7 +31,7 @@ BinaryFile::BinaryFile(const Path& path)
 	file.read(_data.GetPointer<char>(), this->GetSize());
 
 	// Make sure the contents were read correctly
-	if (file.gcount() != this->GetSize())
+	if (static_cast<std::size_t>(file.gcount()) != this->GetSize())
 	{
 		Console::WriteLine("An error occurred while reading the binary file '@'.", path);
 	}
